// src/app/tasks/interfaces/tasks.interface.ts
// Generated by https://quicktype.io


export interface TaskData {
    _id: string;
    title: string;
    description: string;
    assignedTo: string;
    status: string;
    dueDate: string;
    createdAt: string;
    updatedAt: string;
    __v: number;
}

export interface CreateTask{
    title: string,
    description: string,
    status: string,
    dueDate: string
}

export interface UpdateTask extends Partial<CreateTask>{}